+++ KParfume-BackEnd/KParfume.Core/Domain/StavkaKupovine.cs
@@ -14,9 +14,9 @@ namespace KParfume.Core.Domain
         public virtual Parfem Parfem { get; private set; }
         public long sk_kup_id { get; private set; }
         public virtual Kupovina Kupovina { get; private set; }
-        public double sk_kolicina { get; private set; }
+        public int sk_kolicina { get; private set; }
 
-        public StavkaKupovine(double sk_cena_pj, long sk_par_id, long sk_kup_id, double sk_kolicina)
+        public StavkaKupovine(double sk_cena_pj, long sk_par_id, long sk_kup_id, int sk_kolicina)
         {
             this.sk_cena_pj = sk_cena_pj;
             this.sk_par_id = sk_par_id;
